build: Reduce the volume of output printed during the build
Some stages of the build can produce an enormous volume of
output, which makes it hard to see what's going on.   This
change causes make to echo just the name of the command it
is running and the target it is building.   The commands
will still write error messages to standard output if they
fail.

Signed-off-by: Euan Harris <euan.harris@citrix.com>