makedeb: Extract additional information from spec file
The Python RPM library doesn't give us
all the information we need from a spec file, such as
the list of file contents, so we must read these directly
from the spec file.    Don't really want to write a full
spec file parser, but this may be necessary to deal with
conditionals in more complex files.

Signed-off-by: Euan Harris <euan.harris@citrix.com>